[CI skip] Restructuring test classes
