+++ src/test/java/io/github/thebusybiscuit/slimefun4/testing/tests/items/backpacks/TestEnderBackpack.java
@@ -1,8 +1,7 @@
-package io.github.thebusybiscuit.slimefun4.testing.tests.items.implementations.tools;
+package io.github.thebusybiscuit.slimefun4.testing.tests.items.backpacks;
 
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
-import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -13,13 +12,13 @@ import org.junit.jupiter.api.Test;
 import be.seeseemelk.mockbukkit.MockBukkit;
 import be.seeseemelk.mockbukkit.ServerMock;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
-import io.github.thebusybiscuit.slimefun4.implementation.items.tools.PortableDustbin;
+import io.github.thebusybiscuit.slimefun4.implementation.items.backpacks.EnderBackpack;
 import io.github.thebusybiscuit.slimefun4.testing.TestUtilities;
 import io.github.thebusybiscuit.slimefun4.testing.interfaces.SlimefunItemTest;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
-class TestPortableDustbin implements SlimefunItemTest<PortableDustbin> {
+class TestEnderBackpack implements SlimefunItemTest<EnderBackpack> {
 
     private static ServerMock server;
     private static SlimefunPlugin plugin;
@@ -36,24 +35,23 @@ class TestPortableDustbin implements SlimefunItemTest<PortableDustbin> {
     }
 
     @Override
-    public PortableDustbin registerSlimefunItem(SlimefunPlugin plugin, String id) {
-        SlimefunItemStack item = new SlimefunItemStack(id, Material.BUCKET, "&4Test Dustbin");
-        PortableDustbin dustbin = new PortableDustbin(TestUtilities.getCategory(plugin, "dustbin"), item, RecipeType.NULL, new ItemStack[9]);
-        dustbin.register(plugin);
-        return dustbin;
+    public EnderBackpack registerSlimefunItem(SlimefunPlugin plugin, String id) {
+        SlimefunItemStack item = new SlimefunItemStack(id, Material.ENDER_CHEST, "&5Test Ender Backpack");
+        EnderBackpack backpack = new EnderBackpack(TestUtilities.getCategory(plugin, "ender_backpack"), item, RecipeType.NULL, new ItemStack[9]);
+        backpack.register(plugin);
+        return backpack;
     }
 
     @Test
-    @DisplayName("Test Dustbin opening an empty Inventory")
+    @DisplayName("Test Ender Backpack opening Enderchest")
     void testRightClickBehaviour() {
         Player player = server.addPlayer();
-        PortableDustbin dustbin = registerSlimefunItem(plugin, "TEST_PORTABLE_DUSTBIN");
+        EnderBackpack backpack = registerSlimefunItem(plugin, "TEST_ENDER_BACKPACK");
 
-        simulateRightClick(player, dustbin);
+        simulateRightClick(player, backpack);
 
-        // We expect an empty Inventory to be open now
-        Inventory openInventory = player.getOpenInventory().getTopInventory();
-        Assertions.assertTrue(openInventory.isEmpty());
+        // We expect the Enderchest to be open now
+        Assertions.assertEquals(player.getEnderChest(), player.getOpenInventory().getTopInventory());
     }
 
 }
